requirements.txt
furl
tenacity
schema
hydra-core
pydoll-python
tortoise-orm[asyncmy]
cashews[diskcache]
pymongo
dnspython
certifi
orjson
loguru
google-genai
arrow
crawlee
python-dotenv
invoke
